find and apply discount for product in cart

+++ mvc-spring-boot/src/main/java/com/beetech/mvcspringboot/controller/publics/cart/dto/CartItem.java
@@ -17,7 +17,7 @@ public class CartItem {
     @Min(1)
     private Long quantity;
 
-    public Long getTotal() {
+    public Double getTotal() {
         return product.getPrice() * quantity;
     }
 }
